requirements.txt
# Core Dependencies
torch>=1.12.0
torchvision>=0.13.0
nibabel>=3.2.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0

# Medical Imaging
monai>=1.0.0
SimpleITK>=2.2.0
einops>=0.7.0

# Visualization and Logging
matplotlib>=3.5.0
tensorboard>=2.10.0
tqdm>=4.64.0

# Configuration
pyyaml>=6.0

# Utilities
pandas>=1.3.0

